feat: make a new migration with progress_percentage in part

+++ app/models/material.py
@@ -29,7 +29,7 @@ class MaterialPart(BaseModel):
         Enum(MaterialStatusEnum), nullable=False
     )
     expected_delivery_date: Mapped[datetime] = mapped_column(nullable=True)
-    quantity: Mapped[int] = mapped_column(Integer, default=1, nullable=False)
+    quantity: Mapped[float] = mapped_column(Float, default=1, nullable=False)
 
     material: Mapped['Material'] = relationship(back_populates='part_associations')
     part: Mapped['Part'] = relationship(back_populates='material_associations')
